DB 호출 관련 코드 최적화

+++ src/main/java/com/fastcampus/project/sns/controller/PostController.java
@@ -72,7 +72,7 @@ public class PostController {
     }
 
     @GetMapping("/{postId}/likes")
-    public Response<Integer> likeCount(
+    public Response<Long> likeCount(
             @PathVariable Integer postId
     ) {
         return Response.success(postService.likeCount(postId));
